fix(core): use startNetwork retry helper in all core tests
PR #4790 added port-in-use retry logic to createCoreFetcher but missed
all other direct NewNetwork+Start call sites. Extract the retry logic
into a shared startNetwork helper and use it in all core test files.

Affected tests:
- TestBlockFetcherHeaderValues (fetcher_no_race_test.go)
- TestBlockFetcher_GetBlock_and_SubscribeNewBlockEvent (fetcher_test.go)
- TestFetcher_Resubscription (fetcher_test.go)
- TestMakeExtendedHeaderForEmptyBlock (header_test.go)
- TestRemoteClient_Status (client_test.go)
- TestRemoteClient_StartBlockSubscription_And_GetBlock (client_test.go)
