+++ core/fetcher_test.go
@@ -14,8 +14,8 @@ func TestBlockFetcher_GetBlock_and_SubscribeNewBlockEvent(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*15)
 	t.Cleanup(cancel)
 
-	network := NewNetwork(t, DefaultTestConfig())
-	require.NoError(t, network.Start())
+	cfg := DefaultTestConfig()
+	network := startNetwork(t, cfg)
 	t.Cleanup(func() {
 		require.NoError(t, network.Stop())
 	})
@@ -50,8 +50,7 @@ func TestFetcher_Resubscription(t *testing.T) {
 	t.Cleanup(cancel)
 	// run new consensus node
 	cfg := DefaultTestConfig()
-	tn := NewNetwork(t, cfg)
-	require.NoError(t, tn.Start())
+	tn := startNetwork(t, cfg)
 	host, port, err := net.SplitHostPort(tn.GRPCClient.Target())
 	require.NoError(t, err)
 	client := newTestClient(t, host, port)
@@ -86,8 +85,7 @@ func TestFetcher_Resubscription(t *testing.T) {
 
 	// start new consensus node(some components in app can't be restarted)
 	// on the same address and listen for the new blocks
-	tn = NewNetwork(t, cfg)
-	require.NoError(t, tn.Start())
+	tn = startNetwork(t, cfg)
 	t.Cleanup(func() {
 		require.NoError(t, tn.Stop())
 	})
